Add producer support for binary, fix plaintext
This relies on a ugly shortcut, basically whoever implements the
producers must take care of implementing the generateOuput*Message
methods that they will need. This is because the "standard"
generateOutputJsonMessage puts out a string containing (supposedly) a
JSON instead of a proper JSON representation, and going down the "parse
the JSON and extract stuff form that" route for plaintext and binary
would have taken too long.