[#277] GDPR keys to delete should have a correlation id
A GDPR deletion job is run using a list of keys to be deleted. In the case when the output result for multiple keys should be aggregated, this is not easily achievable. This can be solved by passing a correlation id together with the key to deleted, so multiple keys (even different keys for different datastores) can refer to a single logical group.

The correlation id should be placed in a column of the input model, and can be configured inside the DataStore conf declared inside the BatchGdprETLModel.

Correlation id should also be configured when the keys to delete are passed through rest config.