[#225] Embed folder compaction utility
Implement `FolderCompactionUtils` which, driven by configuration, is able to read and rewrite given partitions of an hdfs folder with a new parallelism.

The input configuration looks like the following:

```
{
  "inputModel" : "name of the input model" | ModelConf,
  "outputModel" : "name of the input model" | ModelConf,
  "partitions" : {
    "columnName1" : ["value1", "value2"],
    "columnName2" : ["value1", "value2"]
   },
   "numPartitions" : "integer"
}
```

ModelConf has the following structure:

```
  "name": "a name you like"
  "uri": "URI of the dataset: basePath if the dataset is partitioned"
  "schema": "Spark json representation of the schema"
  "options": {
    saveMode: "spark save mode"
    format: "spark data format"
    extraOptions: {  // "extra format to the spark reader/writer"
      key: value
    }
    partitionBy: [ // "partition columns"
      partitionColumn1,
      partitionColumn2,
      ...
      partitionColumnN
    ]
  }
```

The utility will retrieve the two indicated models and check that they are file-based, it will also check that the indicated columns are partition columns.
If the previous requirements are met then it will generate each combination of column values, sort them alphabetically and then perform read, repartition to the "numPartitions" number and then write to the outputModel.

Ideally it should read one "partition" at a time, repartition to the number of partitions indicated in the conf and then write.