[#279] Implement KafkaBacklogAnalyzer
# New features and improvements

A new abstract producer that is able to check the backlog size of an Spark Structured Streaming ETL.

# Kafka metrics producers

We need to produce a suite of actors able to monitor the metrics of Kafka topics, there are mainly two needs that must be addressed:

- backlog of a list of pipegraphs or etls
- traffic of a list of topics

## Backlog size calculator

In order to calculate the backlog of an ETL, the following information are needed:

- the latest offsets consumed by an ETL (provided by TelemetryActor)
- the latest offsets available in a Kafka topic (provided by KafkaActor)

## Traffic calculator

In order to calculate the traffic of a kafka topic the following information are needed:

- the latest offsets available in a Kafka topic at the beginning of the time window (provided by KafkaActor)
- the latest offsets available in a Kafka topic at the end of the time window (provided by KafkaActor)

## Kafka Actor

Actor tied with a single kafka topic which has the only duty to return the current latest offsets of a Kafka Topic.# Breaking changes

None.

# Migration

None.

# Bug fixes

None.

# Related issue

Closes #279
Closes #236