[#288] Remove the detail of all keys to be deleted from HDFS GDPR deletion
# Scenario summary

As of 2.23.0, the HDFS GDPR deletion handler lists all the keys that will be deleted and outputs them in the final result of the deletion.

# Problem statement

In case of a long list of keys to be deleted, this list could be too large to be held in memory.

# Proposed solution

The outcome of an HDFS deletion should only list the key that should be matched and the type of matching.

**Bonus:** the matching is now done through broadcasting explicitly the input keys, therefore it creates simpler sql statements