[#339] Fix and implement api for pipegraph editing via web ui
# New features and improvements

Implement an endpoint to add a new pipegraph thru UI. Adds additional DTOs to represent pipegraph parts and errors.

# Breaking changes

Implements `/editor/pipegraph` endpoint, described in `wasp-openapi.yaml`. By posting a pipegraph in a described JSON format, you are able to validate and add it to the wasp database. In the case of errors, the backend will return you a list of errors.

# Migration

None

# Bug fixes

None

# Related issue

Closes #339