[#404] Improve whitelabel
# New features and improvements

Introduction of an IoT demo pipegraph.
IoT Industrial Plant pipegraph is modelling a manufacturing use case, simulating domain KPI like OEE, OLE, etc.

A random Producer is generating data on a kafka topic and then we have a simple pipegraph persisting such events on Solr.
In addition there is also an event rule definition to generate events when OEE kpi is lower than 80.

# Breaking changes

# Migration

# Bug fixes

removing typesafe resolver because it is not available anymore

# Related issue

Closes #404