[#424] Support deserialization of Avro enums into strings
# New features and improvements

AvroDeserializerExpression supports deserialization of enums into strings. This means that we are able to deserialise Avro messages which contain enums and they will be mapped to a Spark String (because Spark lacks the enum type). Before this patch it was not possible to deserialise Avro data which contained an enum, because it would've failed at runtime.

# Breaking changes

None.

# Migration

None.

# Bug fixes

None.

# Related issue

Closes #424