[#436] Refactor plugin in order to allow plugin defined models
# New features and improvements

Wasp plugin architecture has been refactored, in order to being capable of add new plugins without the need of add a correspinding model in wasp core module.

Traits DatastoreCategory and DatastoreProduct has been removed, and the new base component of a model is the GenericProduct case class. Every existing product, used by existing plugins, is an instance of the GenericProduct which exposes his category name and product name.

Every plugins that already exists will continue to work as usual, and the new plugins should use the GenericModel, which uses the new GenericProduct.

The GenericModel is structured as following:

    case class GenericModel(override val name: String,
	  kind: String,
	  value: BsonDocument,
	  options: GenericOptions = GenericOptions.default
	  )
	 extends DatastoreModel {
	 override def datastoreProduct: DatastoreProduct = genericProduct
	}

with parameters:

 - name: Model name
 - kind: It could be used whithin a plugin in order to check if the correct model is using
 - value: a BsonDocument with the infos that a further model (plugin specific) could parse
 - options: additional options that might be usefull to the custom plugin

A new custom plugin named **Plugin-parallel-write-spark** has been introduced, which uses the GenericModel above mentioned. It uses a more specific model (ParallelWriteModel), which parses informations that are contained in the GenericModel to use them in the proper way.
The Plugin-parallel-write-spark implements a new **ParallelWriteSparkStructuredStreamingWriter** which, for each microbatch of the stream, makes an http call to a REST API in order to retrieve the s3 bucket path in wich save the microbatch content, and the temporary credentials necessary to write onto it

# Breaking changes

None.

# Migration

None.

# Bug fixes

None.

# Related issue

Closes #436