[#430] Wasp kafka-spark-sql integration is old and caused trouble
# New features and improvements

Add two new modules:

- `spark-sql-kafka-0-11-new`
- `plugin-kafka-spark-new`

These new modules are (another) wasp<->kafka integration, that is most updated with regard to Spark 2.4 branch. We cannot use upstream version because it does not support Kafka header, which we want to support.

To do a recap:

## `spark-sql-kafka-0-11`, `plugin-kafka-spark`

"Current" implementation that was tested in production with the event-engine enabled.

## `spark-sql-kafka-0-11_old`, `plugin-kafka-spark_old`

"Old" implementation that has problems with the event-engine, but performs better than the "current" implementation when stressed with a lot of concurrent streaming queries.

## `spark-sql-kafka-0-11_old`, `plugin-kafka-spark_old`

Newest implementation that has never been tested thoroughly in production but is 1:1 with the upstream Spark one which surely solves the event engine concurrency issue and *should* solve also the contention problem of the current version.

# Breaking changes

None.

# Migration

Nothing, if you want you can try it out changing your dependency to the one with _new suffix.

# Bug fixes

Among others that I cannot find right now:

- SPARK-32044
- SPARK-34187
- SPARK-23623
- SPARK-33962

Still SPARK-32032 is not fixed, because the change was quite big.

# Related issue

Closes #430