[#453] HttpWaspWriter should better handle json serialization
[#453] HttpWaspWriter should handle json serialization better

# New features and improvements

Now the HttpModel has also a structured field which indicates how the request
body will be rendered. The effect of this configuration has effect only if the
DataFrame contains only one column to be sent and only if it is of ArrayType or MapType.
If structured is true the array or map will always be enclosed in a
json object, otherwise the map or the array will be at the top level of the json document.
Input dataframe:

```
+---------+
|  values |
+---------+
|[3, 4, 5]|
+---------+
```

Request with structured = true

```
{"values" : [3, 4, 5]}
```

Request with structured = false

```
[3, 4, 5]
```

# Bug fixes

None.

# Related issue

Closes #453