[#454] Make easier to start wasp in a jvm process
# New features and improvements

It's much easier to start a Single Node WASP cluster.
Running `whitelabel/docker/cdh6/start-wasp-one.sh` will start a docker container with all the needed services and consumers-spark-streaming, master and producers wasp roles in a single JVM. This does not have direct utility, but will serve as a blueprint for vertical applications that will like to do so (for easier operationalisation and savings).

# Breaking changes

None.

# Migration

None.

# Bug fixes

None.

# Related issue

Closes #454