[#446] Create Golden Gate Adapter for AutoDataLake
# New features and improvements

### Purpose of the feature

The following feature has been developed to translate the mutations that have the custom Goldengate format into the format required by the cdc-plugin !325.

NB: Oracle has two types of mutation structure [operation formatter vs row formatter](https://docs.oracle.com/goldengate/bd1221/gg-bd/GADBD/GUID-F0FA2781-0802-4530-B1F0-5E102B982EC0.htm#GADBD481).
- the format generated by the **row formatter**. Where in the schema is represented only the final state of the row without struct field but having all the fields of the table along with the required oracle fields.
- the format generated by the **operation formatter**. The operation formatter instead of having all the fields in the dataframe, has the mandatory GoldenGate fields as fields and has two additional struct field named **before** and **after** that contains the status of the row before and after the application of the mutation in the database.

In this mr only a strategy for the first type of formatter (the row formatter) is provided. We will refer to the schema generated by the row formatter as **flat schema** and we will refer to the schema created by the operation formatter as **nested schema**. So, in a future release #466, we will provide the `GoldenGateAdapterNeastedModelStrategy` while in this release there is only the adapter for the row model implemented by the class `GoldenGateAdapterFlatModelStrategy`.

If you see the [documentation](https://docs.oracle.com/goldengate/bd1221/gg-bd/GADBD/GUID-F0FA2781-0802-4530-B1F0-5E102B982EC0.htm#GADBD526), for the JSON messages they use a before and an after field containing all the data related to the mutation.
Instead, for the Avro examples, they use a schema containing all the data at the same level without nesting the fields in the before and after. This mr has been developed taking the Avro messages and the Avro schema structure as input for the mapping. There will be a future mr that treat the structure where the fields related to the row inside the before and after fields.

The strategy developed applies the following transformation:

it takes as input a dataframe where each row represented in a JSON format has the following shape (where the fields other than "table", "op_type", "op_ts", "current_ts", "pos", "primary_keys" and "tokens" are dependentent to which the mutation refers):

```json
{
    "table": "GG.TCUSTORD",
    "op_type": "I",
    "op_ts": "2013-06-02 22:14:36.000000",
    "current_ts": "2015-09-18T10:13:11.172000",
    "pos": "00000000000000001444",
    "primary_keys": [
        "CUST_CODE",
        "ORDER_DATE",
        "PRODUCT_CODE",
        "ORDER_ID"
    ],
    "tokens": {
        "R": "AADPkvAAEAAEqL2AAA"
    },
    "CUST_CODE": "WILL",
    "ORDER_DATE": "1994-09-30:15:33:00",
    "PRODUCT_CODE": "CAR",
    "ORDER_ID": "144",
    "PRODUCT_PRICE": 17520.0,
    "PRODUCT_AMOUNT": 3.0,
    "TRANSACTION_ID": "100"
}
```

and convert the dataframe to have the following shape:

```json
{
    "key": {
        "CUST_CODE": "WILL",
        "ORDER_DATE": "1994-09-30:15:33:00",
        "PRODUCT_CODE": "CAR",
        "ORDER_ID": "144"
    },
    "value": {
        "timestamp": "2013-06-02 22:14:36.000000",
        "commitId": "00000000000000001444",
        "beforeImage": null,
        "afterImage": {
            "CUST_CODE": "WILL",
            "ORDER_DATE": "1994-09-30:15:33:00",
            "PRODUCT_CODE": "CAR",
            "ORDER_ID": "144",
            "PRODUCT_PRICE": 17520.0,
            "PRODUCT_AMOUNT": 3.0,
            "TRANSACTION_ID": "100"
        },
        "type": "insert"
    }
}
```

This enables the strategy to convert the Goldengate format (supposing that the reader independently
from the format of the input (as for eg JSON, Avro etc..) can translate it to a Spark dataframe with the required shape).
Even if it is not represented in the example, the strategy performs the mapping also for the delete and the update.
The only mutations it doesn't deal with are the truncate operations (because the cdc-writer needs the primary key list of the involved rows, but GoldenGate doesn't provide this information in the mutation).

### Available customizations

If you need to customize the input dataframe incoming into the strategy you can extend the strategy and implement the method: `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.GoldenGateAdapterStrategy#prepareInitialDf `.
You can also customize the dataframe that exit from the GoldenGateAdapterStrategy by override the method `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.GoldenGateAdapterStrategy#enrichFinalDf `.

### Advanced usage/additional features offered by this working item

With the introduction of this feature there are also other utility under the object `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.OrderTableMutation `. Some of the most noticeable are the following methods:
- `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.GoldenGateMutationUtils#extractTableDataset ` that can be used to retrieve the dataset representing the rows of the mutations given as input the dataframe containing the mutations.
- `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.GoldenGateMutationUtils#mapIntoCaseClass ` that can be used to retrieve a dataset of parametric object called `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.OrderTableMutation ` that is a functor that represent the mutation. It contains the methods `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.OrderTableMutation#extractInnerModel ` that enable to retrieve the table representing the mutation. It also offers the well known `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.OrderTableMutation#map ` and `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.OrderTableMutation#flatMap`.

### Usage

For example of usage check the test under `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.GoldenGateAdapterStrategyTest`

# Breaking changes

The structure of the CDC dependent plugins has slightly changed. Has been updated and externalized a trait that implement the concept of field translation under `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.CdcMapper`. If you need to write a new module that translate the mutations writing on delta using the CDC plugin you are required to implement `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.CdcMapper`. This is a reasonable interface for enabling a common language and the reuse of some concept while translating a mutation from a domain to the domain used by the CDC writer. I strongly suggest also ensuring by a unit test that the output format of the translated rows is the same that the CDC plugin expects.

Another change is the moving of the object `CdcModel` that from now on is under `it.agilelab.bigdata.wasp.models.CdcModel`.

# Migration

If you had used the `DebeziumMutationStrategy` in the past the new location is under: `it.agilelab.bigdata.wasp.consumers.spark.strategies.cdc.DebeziumMutationStrategy`.

# Bug fixes

NA.

# Related issue

Changes involves also the feature: !325

Closes #446