[#484] Choose a json framework and stick to it
# New features and improvements

Change every JSON library that is not spray-json to spray-json

Replace json4s in core, consumers-spark, plugin-hbase-spark, plugin-solr-spark and whitelabel with spray-json.

Json4s can't be removed from every module. Nifi-client uses it internally, so json4s in it and in every module that is dependent on it will remain unchanged. Those modules include master, model and openapi. Note that json4s source code in those modules that is not dependent on nifi-client will be changed to spray-json.

Replace lift-json with spray-json in microservice-catalog and plugin-parallel-write-spark modules.


# Breaking changes

None

# Migration

None

# Bug fixes

None

# Related issue

Closes #484