[#505] HttpWriter throws unhelpful ClassCastException when headers column is not a Map
# New features and improvements

HttpWriter accepts also arrays of struct as headers, 
instead of only maps, and does so by casting the first column 
as a string and using it as header key, casting the second column
as a string too and using it as header value, if the struct does not contain
two fields but just one, it will fail. It will issue a warning
when doing so because it might be a coding/configuration mistake and might
not always be safe.

# Breaking changes

None.

# Migration

None.

# Bug fixes

HttpWriter errors are more descriptive.

# Related issue

Closes #505