[#552] Allow to specify mongo authentication database different from data database
# New features and improvements

Now the user can specify an authentication database to mongo that is different from the data database
through the configuration key `wasp.mongo.authentication-db`. Default value is `${wasp.mongo.db-name}`
so that the default behavior is consistent.

# Breaking changes

Potentially binary incompatible change: `ConfigManager` was a Scala `object`, making it hard to test dynamically injected configuration values. I split it into a `class` (with the whole implementation) and an `object` extending the `class`. In this way all usages do not need any source change.

# Migration

None.

# Bug fixes

None.

# How this feature was tested

`ConfigManagerSpec` and existing tests.

# Related issue

Closes #552