[#558] feature: CDP release is affected by MavenGate
# New features and improvements

Wasp does not import any dependency that is affected by [MavenGate](https://blog.oversecured.com/Introducing-MavenGate-a-supply-chain-attack-method-for-Java-and-Android-applications).

# Breaking changes

Wasp CDP build, does not declare as a direct dependency the following:

- org.sklsft.commons:commons-crypto:1.0.0
- org.spire-math:spire_2.11:0.13.0
- org.spire-math:spire-macros_2.11:0.13.0
- net.jpountz.lz4:lz4:1.2.0

Users that were relying on these dependencies should provide them in a different way.

# Migration

Users should add the afore-mentioned dependencies to Spark driver and executor classpaths in a way different than relying on sbt packaging, for example through Cloudera parcels.

# Bug fixes

None.

# How this feature was tested

Existing tests and a one customer project.

# Related issue

Closes #558