[#556] Add permissive mode of parsing to WBDS
# New features and improvements

Sometimes reading from a source with serialized data in avro/json, can happen that a record with some typo or serialized with an unknown avro schema came in, with the actual readers behaviour, when this happens, an exceptions will be raised and pipegraph will crash. The decision on how to handle parsing errors should be left to downstream implementations, especially for streaming jobs in which is often desirable that the application keep running.

The way to change parsing modes is to add a `parsingMode` option in the reader options of your strategy.

An example:

```scala
val streamingInput = StreamingReaderModel.kafkaReader(
  topicModel.name,
  topicModel,
  rateLimit,
  Map("parsingMode" -> "handle")
)
```

Avalaible options are:

- **strict**
  - throws an exception if there is an malformed row(default behaviour)
- **handle**
  - returns the malformed rows to the user, so that the downstream can handle them as preferred
- **ignore**
  - filters out malformed records and continues job execution without them

NOTE: This feature is supported only for kafkaReader.

# Breaking changes

None.

# Migration

None.

# Bug fixes

None.

# How this feature was tested

Unit tests -\> KafkaSparkStructuredStreamingReaderSpec.scala

# Related issue

Closes #556