[!449] - Resolve "Implement a streaming query watchdog"
# New features and improvements

A streaming query will be automatically killed if it exceeds a time bound of `triggerInterval * multiplier`.

The configuration of the multiplier is global for now.

Relevant configurations are:

```
wasp {
  spark-streaming {
    watchdog {
      trigger-interval-multiplier = 20
      enabled = false
    }
  }
}
```

`trigger-interval-multiplier`: a double that will be used to determine the deadline to kill a streaming query via triggerInterval * multiplier

`enabled`: a boolean it is disabled by default, set to true to enable

# Breaking changes

None.

# Migration

None.

# Bug fixes

None.

# How this feature was tested

unit tests: `TestStuckQueryStrategy.scala`

# Related issue

Closes #568