[!446] - Hotfix: Json Parsing Mode not-working-correctly
# New features and improvements

None.

# Breaking changes

None.

# Migration

None.

# Bug fixes

Strict handling mode for parsing Json was not working properly, and it would still parse and validate records with schema not matching the model of ETL.
An example would be that a :
```
{"value": "sss"}
``` 

would pass for a model that had a definition of:
```
case class(value: String, value2: String)
```
where the value2 is non-nullable. If a field can be nullable, it should be defined in the model as `Option[T]`.

# How this feature was tested

`KafkaStructuredStreamingReaderSpec`

# Related issue

Closes #567