[!471] - Resolve "Kafka headers as optional field"
# New features and improvements

None.

# Breaking changes

None.

# Migration

None.

# Bug fixes

If the user does not specify kafka configuration "includeHeaders" (either reader or global kafka) the reading will fail as follows:

```
Caused by: org.apache.spark.sql.AnalysisException: Column 'headers' does not exist. Did you mean one of the following? [key, offset, raw, value, topic, timestamp, partition, timestampType, computedValue];
```

Given that the overhead of adding the header to the dataframe is very low, we do not let the user "exclude" the headers via setting `includeHeaders=false`, then the header configuration of the topic will be used to either include it or not in the dataframe returned to the user (i.e. to `Strategy#transform`). If the users sets it a warn will be issued and then overwritten.

# How this feature was tested

Add unit test in `KafkaSparkStructuredStreamingReaderSpec`.

# Related issue

Closes #586